refactor: GameController 클래스 안에서 Car 객체 생성
